requirements.txt
requests
python-dotenv
halo
csv